python/pyproject.toml
[project]
name = "agent-manager"
version = "0.2.0"
description = "Workflow orchestration system for local AI agents"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Agent Manager Contributors"}
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.9.0",
    "rich>=13.0",  # For nice CLI output
    "eval-type-backport>=0.2.0;python_version<'3.10'",  # For type annotation compatibility
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
    "mypy>=1.0",
]
# Future provider dependencies
anthropic = ["anthropic>=0.18"]
ollama = ["ollama>=0.1"]
openai = ["openai>=1.0"]

[project.scripts]
agentctl = "agentctl:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_manager", "agentctl"]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.9"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
